requirements.txt
appnope==0.1.0
astroid==1.6.2
Babel==2.6.0
backcall==0.1.0
bleach==2.1.3
boto3==1.7.46
botocore==1.10.46
certifi==2018.4.16
chardet==3.0.4
click==6.7
coverage==4.5.1
cycler==0.10.0
decorator==4.3.0
dj-database-url==0.5.0
Django==1.11
django-forms-bootstrap==3.1.0
django-localflavor==2.0
django-phonenumber-field==2.0.0
django-storages==1.6.6
docutils==0.14
entrypoints==0.2.3
flake8==3.5.0
Flask==0.12.2
flask-paginate==0.5.1
Flask-PyMongo==0.5.1
gunicorn==19.8.1
html5lib==1.0.1
idna==2.6
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.1
numpy==1.14.2
oauthlib==2.0.7
pandas==0.22.0
pandocfilters==1.4.2
parso==0.2.0
pbr==4.0.4
pexpect==4.5.0
phonenumberslite==8.9.9
pickleshare==0.7.4
Pillow==5.1.0
prettytable==0.7.2
prompt-toolkit==1.0.15
psycopg2==2.7.4
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.3
pylint-plugin-utils==0.3
pymongo==3.6.1
PyMySQL==0.8.0
pyparsing==2.2.0
pyPdf==1.13
PySocks==1.6.8
python-dateutil==2.7.2
pytz==2018.4
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.13
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
stevedore==1.28.0
stripe==1.82.2
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
tweepy==3.6.0
urllib3==1.22
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==3.3.1
widgetsnbextension==3.2.1
wrapt==1.10.11
